Add OpenCR config update
OpenCR depends on files (or Env Vars) for configuration. To configure
OpenCR for this usecase we need a new config file that would overwrite
the generic data from the default client package settings.
Environment variables are not a viable option here as that would
require recreating the OpenCR container (instead of just restarting it)

By copying in the new file into the volumed directory the config will be
persisted and the container can just be restarted to get the new
settings.

DEB-20