Use axios to send resources to FHIR server
This commit refactor the resource download function such that
it now uses the axios npm package to send resources to the
FHIR server.

This change was necessary for reducing the complexity of the
fhirConfig.js script. Due the introduction of this package, a couple of
environmental variables are no longer supported. The script now
supports the following variables:
 - `HAPI_FHIR_BASE_URL`: the full base url for the HAPI FHIR server (e.g
 `http://localhost:3447/fhir)
 - `COVID19_IG_URL`: the full url to the COVID19 implementation guide.

Related Tickets:
 - DEB-27
 - DEB-8
 - DSC19-75